+++ packages/atomic-angular/projects/atomic-angular/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.30](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@2.3.29...@coveo/atomic-angular@2.3.30) (2023-02-21)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+## [2.3.29](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@2.3.28...@coveo/atomic-angular@2.3.29) (2023-02-20)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
 ## [2.3.28](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@2.3.27...@coveo/atomic-angular@2.3.28) (2023-02-15)
 
 **Note:** Version bump only for package @coveo/atomic-angular